Refactor and addition of archive users command in teams command
The teams commands is refactored. The users management for teams
has been moved to a subcommand "users" in order to:

- Include another archive command
- Provide more clarity
