+++ commands/command.go
@@ -40,7 +40,7 @@ var CommandDeleteCmd = &cobra.Command{
 	Short:      "Delete a slash command",
 	Long:       `Delete a slash command. Commands can be specified by command ID.`,
 	Example:    `  command delete commandID`,
-	Deprecated: "This command is deprecated, please use archive instead",
+	Deprecated: "please use \"archive\" instead",
 	Args:       cobra.ExactArgs(1),
 	RunE:       withClient(archiveCommandCmdF),
 }
